requirements.txt
asgiref==3.8.1
Django==5.0.4
django-constance==3.1.0
django-cors-headers==4.3.1
django-picklefield==3.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
inflection==0.5.1
packaging==24.0
PyJWT==2.8.0
pytz==2024.1
PyYAML==6.0.1
redis==5.0.3
sqlparse==0.4.4
uritemplate==4.1.1
